refactored workflow to controllers -> services -> repos from controllers -> repos
