+++ src/accounts/accounts.controller.ts
@@ -5,48 +5,36 @@ import {
   HttpCode,
   HttpStatus,
   Param,
-  NotFoundException,
   Body,
-  BadRequestException,
 } from '@nestjs/common';
-import { AccountsRepository } from './accounts.repository';
-import { TransactionsRepository } from './../transactions/transactions.repository';
 import { AccountDto } from './dto/account.dto';
 import { TransactionDto } from './../transactions/dto/transaction.dto';
+import { AccountsService } from './accounts.service';
 
 @Controller('accounts')
 export class AccountsController {
-  constructor(
-    private readonly accountsRepository: AccountsRepository,
-    private readonly transactionsRepository: TransactionsRepository,
-  ) {}
+  constructor(private readonly accountsService: AccountsService) {}
 
   @Get()
   findAll(): AccountDto[] {
-    return this.accountsRepository.findAll();
+    return this.accountsService.findAll();
   }
 
   @Get(':id')
   findOne(@Param('id') id: string): AccountDto {
-    const account = this.accountsRepository.findById(id);
-
-    if (!account) {
-      throw new NotFoundException(`Account with ID ${id} not found.`);
-    }
-
-    return account;
+    return this.accountsService.findById(id);
   }
 
   @Get(':id/transactions')
   getAccountTransactions(@Param('id') id: string): TransactionDto[] {
-    return this.transactionsRepository.findAllByAccountId(id);
+    return this.accountsService.getAccountTransactions(id);
   }
 
   @Post()
   @HttpCode(HttpStatus.CREATED) // sets res status code to '201 Created' -> customary for successful account creation
   createAccount(@Body() account: AccountDto): AccountDto {
     // @body decorator is used to extract data from req body
-    return this.accountsRepository.create(account);
+    return this.accountsService.createAccount(account);
   }
 
   @Post(':id/transactions/add')
@@ -55,18 +43,7 @@ export class AccountsController {
     @Param('id') accountId: string,
     @Body() transaction: TransactionDto,
   ): TransactionDto {
-    const account = this.accountsRepository.findById(accountId);
-    if (!account) {
-      throw new NotFoundException(`Account with ID ${accountId} not found.`);
-    }
-
-    account.balance.amount += transaction.amount_money.amount;
-
-    const newTransaction = this.transactionsRepository.createTransaction(
-      accountId,
-      transaction,
-    );
-    return newTransaction;
+    return this.accountsService.addTransaction(accountId, transaction);
   }
 
   @Post(':id/transactions/withdraw')
@@ -75,28 +52,7 @@ export class AccountsController {
     @Param('id') accountId: string,
     @Body() transaction: TransactionDto,
   ): TransactionDto {
-    const account = this.accountsRepository.findById(accountId);
-    if (!account) {
-      throw new NotFoundException(`Account with ID ${accountId} not found.`);
-    }
-
-    // validate sender balance
-    if (
-      account.balance.amount === 0 ||
-      account.balance.amount < transaction.amount_money.amount
-    ) {
-      throw new BadRequestException(
-        'Insufficient balance to withdraw the funds.',
-      );
-    }
-
-    account.balance.amount -= transaction.amount_money.amount;
-
-    const newTransaction = this.transactionsRepository.createTransaction(
-      accountId,
-      transaction,
-    );
-    return newTransaction;
+    return this.accountsService.withdrawTransaction(accountId, transaction);
   }
 
   @Post(':id/transactions/send')
@@ -105,46 +61,6 @@ export class AccountsController {
     @Param('id') accountId: string,
     @Body() transaction: TransactionDto,
   ): TransactionDto {
-    // ensure both accounts exist
-    const senderAccount = this.accountsRepository.findById(accountId);
-    if (!senderAccount) {
-      throw new NotFoundException(`Account with ID ${accountId} not found.`);
-    }
-    const targetAccount = this.accountsRepository.findById(
-      transaction.target_account_id,
-    );
-    if (!targetAccount) {
-      throw new NotFoundException(
-        `Account with ID ${transaction.target_account_id} not found.`,
-      );
-    }
-    // validate transaction amount
-    if (
-      transaction.amount_money.amount < 1 ||
-      transaction.amount_money.amount > 1000
-    ) {
-      throw new BadRequestException(
-        'Invalid balance amount: cannot send less than $1 or more than $1000.',
-      );
-    }
-
-    // validate sender balance
-    if (
-      senderAccount.balance.amount === 0 ||
-      senderAccount.balance.amount < transaction.amount_money.amount
-    ) {
-      throw new BadRequestException(
-        'Insufficient balance to make the transfer.',
-      );
-    }
-
-    senderAccount.balance.amount -= transaction.amount_money.amount;
-    targetAccount.balance.amount += transaction.amount_money.amount;
-
-    const newTransaction = this.transactionsRepository.createTransaction(
-      accountId,
-      transaction,
-    );
-    return newTransaction;
+    return this.accountsService.sendTransaction(accountId, transaction);
   }
 }
