+++ src/transactions/transactions.module.ts
@@ -1,7 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TransactionsService } from './transactions.service';
+import { TransactionsRepository } from './transactions.repository';
+import { TransactionsController } from './transactions.controller';
 
 @Module({
-  providers: [TransactionsService],
+  controllers: [TransactionsController],
+  providers: [TransactionsService, TransactionsRepository],
 })
 export class TransactionsModule {}
