+++ src/transactions/transactions.service.ts
@@ -1,4 +1,14 @@
 import { Injectable } from '@nestjs/common';
+import { TransactionsRepository } from './transactions.repository';
+import { TransactionDto } from './dto/transaction.dto';
 
 @Injectable()
-export class TransactionsService {}
+export class TransactionsService {
+  constructor(
+    private readonly transactionsRepository: TransactionsRepository,
+  ) {}
+
+  findAll(): TransactionDto[] {
+    return this.transactionsRepository.findAll();
+  }
+}
